requirements.txt
pywebview
pyinstaller
Pillow
